Add files via upload
A convolutional neural network built to predict whether a patient has no illness, COVID or pneumonia based on their x-ray lung scan. The metrics of the model were then visualised, so that any discrepancies between the metrics for training and validation could be seen. The hyperparameters of the model were then tuned so that the best iteration of the model could be found. The best model was then evaluated using a classification report and confusion matrix.